+++ FlightID.txt
@@ -110,7 +110,6 @@
 0204060,PEK3,0745,0950,HRB,CA1689,738,CY,2:05
 0204060,HRB,1450,1655,PEK3,CA1690,738,CY,2:05
 1030507,PEK3,0740,0945,HRB,CA1603,738,CY,2:05
-1234567,PEK3,0830,1150,ICN,CA123,333,JY,2:05
 1234567,PEK3,1105,1410,ICN,CA131,321,JY,2:05
 1234567,ICN,1510,1615,PEK3,CA132,321,JY,2:05
 1234567,PEK3,1845,2150,GMPI,CA137,738,JY,2:05
@@ -451,7 +450,7 @@
 1234567,PEK3,1340,1805,BKK,CA959,772,JY,5:25
 0204067,NRT1,0850,1320,CTU1,CA460,320,JY,5:30
 1234567,DLC,0900,1430,CTU2,CA8925,738,CY,5:30
-1030560,PEK3,0725,1330,YIN,CA1233,738,CY,6:05
+1030560,PEK,0725,1330,YIN,CA1233,738,CY,6:05
 1234567,PEK3,1800,2050,CTU2,CA4104,321,CY,2:50
 1004000,PEK3,1155,1430,ULN,CA901,738,JY,6:45
 1030560,NRT1,1715,2300,CTU1,CA6701,763,JY,6:45
