+++ control.cpp
@@ -1,0 +1,8 @@
+//#include"mainTemp.cpp"
+//using namespace std;
+//
+////int main()
+////{
+////	cout << sizeof(FlightID) << endl << sizeof(FlightTicket) << endl;
+////	return 0;
+////}
